+++ src/project/diff/DifferenceEngine.ts
@@ -1,4 +1,5 @@
 import * as _ from "lodash";
+import { HandlerContext } from "../../HandlerContext";
 
 import { Chain } from "./Chain";
 
@@ -20,8 +21,8 @@ export class DifferenceEngine {
      * @param baseSha
      * @param headSha
      */
-    public run(baseSha: string, headSha: string) {
-        const baseProjectPromise = this.cloneRepo(this.githubIssueAuth, baseSha);
+    public run(context: HandlerContext, baseSha: string, headSha: string) {
+        const baseProjectPromise = this.cloneRepo(context, this.githubIssueAuth, baseSha);
         baseProjectPromise.then(project => {
             const baseFingerprintPromises: Array<Promise<any>> = _.map(this.chains, c => c.extractor.extract(project));
             Promise.all(baseFingerprintPromises).then(baseFps => {
@@ -38,8 +39,8 @@ export class DifferenceEngine {
         });
     }
 
-    private cloneRepo(githubIssueAuth: GithubIssueAuth, sha: string): Promise<GitProject> {
-        return GitCommandGitProject.cloned(
+    private cloneRepo(context: HandlerContext, githubIssueAuth: GithubIssueAuth, sha: string): Promise<GitProject> {
+        return GitCommandGitProject.cloned(context,
             { token: githubIssueAuth.githubToken },
             new GitHubRepoRef(githubIssueAuth.owner, githubIssueAuth.repo, githubIssueAuth.sha));
     }