+++ src/project/git/GitCommandGitProject.ts
@@ -1,4 +1,5 @@
 import { exec } from "child-process-promise";
+import { HandlerContext } from "../../HandlerContext";
 
 import { isLocalProject, ReleaseFunction } from "../local/LocalProject";
 import { Project } from "../Project";
@@ -58,7 +59,7 @@ export class GitCommandGitProject extends NodeFsLocalProject implements GitProje
      * @param {DirectoryManager} directoryManager
      * @return {Promise<GitCommandGitProject>}
      */
-    public static cloned(credentials: ProjectOperationCredentials,
+    public static cloned(context: HandlerContext, credentials: ProjectOperationCredentials,
                          id: RemoteRepoRef,
                          opts: CloneOptions = DefaultCloneOptions,
                          directoryManager: DirectoryManager = DefaultDirectoryManager): Promise<GitProject> {