+++ test/project/git/GitStatusTest.ts
@@ -1,5 +1,6 @@
 import "mocha";
 import * as assert from "power-assert";
+import { HandlerContext } from "../../../src/HandlerContext";
 
 import * as stringify from "json-stringify-safe";
 
@@ -21,7 +22,7 @@ const Creds = { token: GitHubToken };
 describe("git status analysis", () => {
 
     function freshClone(repoRef: GitHubRepoRef = ExistingRepoRef): Promise<GitProject> {
-        return GitCommandGitProject.cloned(Creds, repoRef, {}, TmpDirectoryManager);
+        return GitCommandGitProject.cloned({} as HandlerContext, Creds, repoRef, {}, TmpDirectoryManager);
     }
 
     it("should recognize a clean repository as clean", done => {